requirements.txt
requests
rsa
debugpy
Pillow
ddddocr==1.5.6
PyQt5
pywin32
nuitka
pycryptodome
keyring